Revert "move flatten fa to its own file"
This reverts commit c2790e0c09929334934bd76bb45bbff66a5e2937.
